image-prompt-app/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
python-multipart
pytest
httpx
Pillow
numpy
scikit-image
open-clip-torch
